Tokens MW tests: token-mw -> tokens-mw, ref tests to use mocks
